Move TaskList.java out of task package
