+++ src/main/java/duke/Duke.java
@@ -1,7 +1,5 @@
 package duke;
 
-import duke.task.TaskList;
-
 import java.io.IOException;
 
 public class Duke {
