+++ src/main/java/duke/Ui.java
@@ -1,6 +1,9 @@
 package duke;
 
 import duke.task.*;
+import duke.testinginterface.DeadlineInterface;
+import duke.testinginterface.EventInterface;
+import duke.testinginterface.TodoInterface;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -17,7 +20,7 @@ public class Ui {
     }
 
     public void showWelcome() {
-        System.out.println("Hello! I'm duke.Duke\nWhat can I do for you?");
+        System.out.println("Hello! I'm Duke\nWhat can I do for you?");
     }
 
     public String readCommand() {
