+++ src/main/java/duke/command/AddCommand.java
@@ -4,7 +4,7 @@ import duke.task.Deadline;
 import duke.task.Event;
 import duke.Storage;
 import duke.task.Task;
-import duke.task.TaskList;
+import duke.TaskList;
 import duke.task.Todo;
 import duke.Ui;
