+++ src/main/java/duke/command/Command.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.*;
-import duke.task.TaskList;
+import duke.TaskList;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException;
