+++ src/main/java/duke/command/DoneCommand.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.Storage;
-import duke.task.TaskList;
+import duke.TaskList;
 import duke.Ui;
 
 public class DoneCommand extends Command {
