+++ src/main/java/duke/command/ListCommand.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.Storage;
-import duke.task.TaskList;
+import duke.TaskList;
 import duke.Ui;
 
 public class ListCommand extends Command {
